Alteração Exercício 5

+++ javaaula2tb/Exercicio5.java
@@ -6,10 +6,10 @@ public class Exercicio5 {
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        float resultado;
+        double resultado;
 
         System.out.println("Qual o comprimento das laterais do seu quadrado?");
-        float lado = sc.nextFloat();
+        double lado = sc.nextDouble();
         System.out.println("O comprimento da lateral e: " + lado + " cm");
         
         //Area = lado x lado
